Add redis docker-compose, refactor hooks
