+++ docker-compose-redis.yaml
@@ -1,0 +1,13 @@
+version: '3.8'
+
+services:
+  redis:
+    image: 'redis:alpine'
+    restart: always
+    ports:
+      - ${REDIS_PORT}:${REDIS_PORT}
+    volumes:
+      - redis-data:/data
+
+volumes:
+  redis-data:
