+++ frontend/src/App.tsx
@@ -18,11 +18,11 @@ import HomePage from './pages/HomePage';
 import ProblemPage from './pages/problems/ProblemPage';
 import QuestionsPage from './pages/QuestionsPage';
 import Loading from './components/Loading';
-import { useDarkMode }  from './components/Navbar/useDarkMode';
+import { useDarkMode } from './hooks/useDarkMode';
 import SocketChatPageTest from './pages/problems/SocketChatPageTest';
 
 function App() {
-  const {theme, toggleTheme, componentMounted} = useDarkMode();
+  const { theme, toggleTheme, componentMounted } = useDarkMode();
   const navigate = useNavigate();
   const queryClient = new QueryClient();
   const [session, setSession] = useState<Session | null>();
